loop over peaks in the inner loop instead of outer loop
It makes more sense to have the harmonic number as the outer loop. This
way, for each harmonic, we look at all peaks. Previously, for each peak,
we checked whether the peak was a possible 1st, 2nd, or 3rd harmonic. A
peak could not possibly be multiple harmonics, so this doesn't make
sense. The inner "peak" loop might be able to be optimized further by
not starting the loop at the beginning if we've already found a harmonic
from the previous outer loop iteration.